fix: comprehensive bug fix - algorithm AM/PM parsing, friend matching, group crash, studentRepo separator, groupsRepo JPQL field, Category sequence conflict, repo ID types

+++ StableRoomie/src/main/java/in/edu/ssn/hostel/repo/categoryRepo.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import in.edu.ssn.hostel.model.Category;
 
 @Repository
-public interface categoryRepo extends JpaRepository<Category, Integer> {
+public interface categoryRepo extends JpaRepository<Category, Long> {
 }
